requirements.txt
# Core dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
python-dotenv>=1.0.0
responses>=0.23.0

# Logging and utilities
structlog>=23.0.0

# Data visualization and analysis
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.5.0
reportlab>=3.6.0
openpyxl>=3.0.0

# Future dependencies (commented for now)
# flask>=2.3.0
# requests>=2.31.0
# streamlit>=1.28.0
